Moved ldtk.rs to its own folder, factored out FieldInstance stuff, added module doc
